+++ geometry_wars/ui/GameObject.h
@@ -6,8 +6,8 @@
 class GameObject {
   public:
     virtual void draw(GameBuffer buffer, Camera* camera) {}
-    virtual bool is_inside(int x, int y, int buffer_width, int buffer_height) { return false; }
-    virtual bool is_outside(int x, int y, int buffer_width, int buffer_height) { return !is_inside(x, y, buffer_width, buffer_height); }
+    virtual bool is_inside(float x, float y, int buffer_width, int buffer_height) { return false; }
+    virtual bool is_outside(float x, float y, int buffer_width, int buffer_height) { return !is_inside(x, y, buffer_width, buffer_height); }
     virtual void mouse_outside(bool lmb_pressed, bool rmb_pressed) {}
     virtual void mouse_press(int button) {}
     virtual void mouse_release(int button) {}
